requirements.txt
paho-mqtt
python-valve